Moved custom button to chainaddresses, and subdomain components
