+++ ui/app/rif/pages/domainsDetailPage/domainDetailActive/domainDetailActive.js
@@ -27,6 +27,11 @@ const styles = {
     noData: '',
     content: '',
     contentActions: '',
+    customButton: {
+      button: '',
+      icon: '',
+      text: '',
+    },
     pagination: {
       body: '',
       buttonBack: '',
@@ -48,6 +53,11 @@ const styles = {
     noData: '',
     content: '',
     contentActions: '',
+    customButton: {
+      button: '',
+      icon: '',
+      text: '',
+    },
     pagination: {
       body: '',
       buttonBack: '',
@@ -69,6 +79,11 @@ const styles = {
     noData: '',
     content: '',
     contentActions: '',
+    customButton: {
+      button: '',
+      icon: '',
+      text: '',
+    },
     pagination: {
       body: '',
       buttonBack: '',
@@ -84,11 +99,9 @@ class DomainsDetailActiveScreen extends Component {
 	static propTypes = {
     addNewChainAddress: PropTypes.func.isRequired,
     setNewResolver: PropTypes.func.isRequired,
-    setChainAddressForResolver: PropTypes.func.isRequired,
     deleteChainAddressForResolver: PropTypes.func.isRequired,
 		setAutoRenew: PropTypes.func.isRequired,
     getUnapprovedTransactions: PropTypes.func,
-    showTransactionConfirmPage: PropTypes.func,
     domain: PropTypes.object.isRequired,
 		domainName: PropTypes.string.isRequired,
 		address: PropTypes.string.isRequired,
@@ -100,9 +113,7 @@ class DomainsDetailActiveScreen extends Component {
 		isOwner: PropTypes.bool,
 		isLuminoNode: PropTypes.bool,
 		isRifStorage: PropTypes.bool,
-    showDomainsDetailPage: PropTypes.func.isRequired,
     displayToast: PropTypes.func.isRequired,
-    waitForListener: PropTypes.func,
     disableResolvers: PropTypes.bool,
     updateChains: PropTypes.bool,
     getDomain: PropTypes.func,
@@ -111,43 +122,14 @@ class DomainsDetailActiveScreen extends Component {
 	constructor (props) {
 		super(props);
     const resolvers = Object.assign([], GET_RESOLVERS());
-    const slipChainAddresses = Object.assign([], SLIP_ADDRESSES);
     const enableComboResolvers = this.props.isOwner && !(props.disableResolvers || false);
 		this.state = {
       disableCombo: !enableComboResolvers,
 			resolvers: resolvers,
-      slipChainAddresses: slipChainAddresses,
-      selectedChainAddress: slipChainAddresses[0].chain,
-			insertedAddress: '',
       updateChains: true,
-      addChainAddress: false,
-      addSubdomain: false,
 		};
 	}
-  updateChainAddress = (selectedOption) => {
-    this.setState({ selectedChainAddress: selectedOption });
-	}
-	updateAddress = (address) => {
-		this.setState({ insertedAddress: address });
-	}
-  async addAddress (e) {
-    const transactionListenerId = await this.props.setChainAddressForResolver(this.props.domainName, this.state.selectedChainAddress, this.state.insertedAddress);
-    this.props.waitForListener(transactionListenerId)
-      .then(transactionReceipt => {
-        this.props.showDomainsDetailPage({updateChains: true, ...this.props.domain});
-      });
-    this.props.showTransactionConfirmPage({
-      action: () => this.props.showDomainsDetailPage(this.props.domain),
-    });
-	}
-
-  showAddChainAddress = () => {
-    this.setState({addChainAddress: !this.state.addChainAddress})
-  }
-  showAddSubdomain = () => {
-    this.setState({addSubdomain: !this.state.addSubdomain})
-  }
-
+  /*
 	async onChangeComboResolvers (e) {
     for (const resolverItem of e.target.children) {
       if (resolverItem.value === e.target.value) {
@@ -177,7 +159,7 @@ class DomainsDetailActiveScreen extends Component {
       }
     }
   }
-
+*/
   componentDidUpdate (prevProps, prevState) {
     if (this.props.disableResolvers !== this.state.disableCombo) {
       this.setState({ disableCombo: this.props.disableResolvers });
@@ -193,7 +175,7 @@ class DomainsDetailActiveScreen extends Component {
   }
 
 	render () {
-    const { domainName, content, expirationDate, autoRenew, ownerAddress, isOwner, isLuminoNode, isRifStorage, selectedResolverAddress } = this.props;
+    const { domain, domainName, content, expirationDate, autoRenew, ownerAddress, isOwner, isLuminoNode, isRifStorage, selectedResolverAddress } = this.props;
     const domainInfo = {
       domainName,
       expirationDate,
@@ -215,58 +197,21 @@ class DomainsDetailActiveScreen extends Component {
           {resolvers &&
           <div id="chainAddressesBody" className={''}>
             <ChainAddresses
+              domain={domain}
               domainName={domainName}
               selectedResolverAddress={selectedResolverAddress}
               paginationSize={3}
               classes={styles.chainAddresses}
               isOwner={isOwner}
             />
-            {(isOwner && resolvers.find(resolver => resolver.address === selectedResolverAddress)) &&
-            <div>
-              <CustomButton
-                svgIcon={SVG_PLUS}
-                text={'Add Address'}
-                onClick={() => this.showAddChainAddress()}
-                className={
-                  {
-                    button: '',
-                    icon: '',
-                    text: '',
-                  }
-                }
-              />
-              {this.state.addChainAddress &&
-                <AddNewChainAddressToResolver
-                  updateChainAddress={this.updateChainAddress.bind(this)}
-                  updateAddress={this.updateAddress.bind(this)}
-                  slipChainAddresses={this.state.slipChainAddresses}
-                  confirmCallback={this.addAddress()}
-                />
-              }
-            </div>
-            }
           </div>
           }
-          <Subdomains domainInfo={domainInfo} classes={styles.subdomains} paginationSize={3}/>
-          {isOwner &&
-            <div>
-              <CustomButton
-                svgIcon={SVG_PLUS}
-                text={'Add Subdomain'}
-                onClick={() => this.showAddSubdomain()}
-                className={
-                  {
-                    button: '',
-                    icon: '',
-                    text: '',
-                  }
-                }
-              />
-              {this.state.addSubdomain &&
-                <AddNewSubdomain />
-              }
-            </div>
-          }
+          <Subdomains
+            isOwner={isOwner}
+            domainInfo={domainInfo}
+            classes={styles.subdomains}
+            paginationSize={3}
+          />
           <LuminoNetworkChannels
             isOwner={isOwner}
             paginationSize={3}
@@ -305,18 +250,9 @@ const mapDispatchToProps = dispatch => {
 	return {
 		addNewChainAddress: (message) => dispatch(rifActions.showModal(message)),
     setNewResolver: (domainName, resolverAddress) => dispatch(rifActions.setResolverAddress(domainName, resolverAddress)),
-    setChainAddressForResolver: (domainName, chain, chainAddress) => dispatch(rifActions.setChainAddressForResolver(domainName, chain, chainAddress)),
     deleteChainAddressForResolver: (domainName, chain, chainAddress) => {},
     getUnapprovedTransactions: () => dispatch(rifActions.getUnapprovedTransactions()),
-    showTransactionConfirmPage: (afterApproval) => dispatch(rifActions.goToConfirmPageForLastTransaction(afterApproval)),
-    waitForListener: (transactionListenerId) => dispatch(rifActions.waitForTransactionListener(transactionListenerId)),
 		setAutoRenew: () => {},
-    showDomainsDetailPage: (props) => dispatch(rifActions.navigateTo(pageNames.rns.domainsDetail, {
-      tabOptions: {
-        screenTitle: 'Domain Details',
-      },
-      ...props,
-    })),
     displayToast: (message) => dispatch(niftyActions.displayToast(message)),
     getDomain: (domainName) => dispatch(rifActions.getDomain(domainName)),
     showToast: (message, success) => dispatch(niftyActions.displayToast(message, success)),
